chore(generate.yml): uncomment Terraform Apply and Terraform State Push steps
chore(generate.yml): uncomment generate-action and output-action steps

The Terraform Apply and Terraform State Push steps were uncommented to enable the execution of Terraform apply and state push commands. This is necessary for the deployment and management of infrastructure resources.

The generate-action and output-action steps were also uncommented to enable the generation of commands and the output of the action edit link. These steps are crucial for the overall workflow and automation process.